Added several activities that house each challenge and the buttons now allow user to switch to a different screen every activity
