+++ app/src/main/java/com/example/a13challenges/CH10.java
@@ -1,0 +1,27 @@
+package com.example.a13challenges;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.EditText;
+import android.widget.TextView;
+
+public class CH10 extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_ch10);
+    }
+    //challenge 10 button function, takes a number and counts from 1 up to the given number
+    public void c10 (View view) {
+        TextView textView = findViewById(R.id.display);
+        EditText numEditText = findViewById(R.id.Num_Ent_1);
+        int num = Integer.parseInt(numEditText.getText().toString());
+        textView.setText(" ");
+        for(int i = 1; i <= num; i++){
+            textView.append(i + " ");
+        }
+    }
+}
