+++ app/src/main/java/com/example/a13challenges/CH8.java
@@ -1,0 +1,25 @@
+package com.example.a13challenges;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.EditText;
+import android.widget.TextView;
+
+public class CH8 extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_ch8);
+    }
+    //challenge 8 button function, takes a number in binary form and converts to decimal
+    public void c8(View view){
+        TextView display = findViewById(R.id.display);
+        EditText num = findViewById(R.id.Num_Ent_1);
+
+        int numOne = Integer.parseInt(num.getText().toString(),2);
+        display.setText(Integer.toString(numOne));
+    }
+}
